Add dummy PIC IRQ support for Simple Vectored Architectures
The current implementation assumes that the architecture supports the Programmable
Interrupt Controller interrupt API in RTEMS. Most architectures do not use this
model and provide only the Simple Vectored interrupt API. The SPARC port has an
implementation of the PIC API that builds on the Simple Vectored API.

The code added here is just enough to compile. The hope is that a generic
PIC API wrapper will be implemented on top of the Simple Vectored which will
instantly provide support for all other architectures.

For now, the issues of code portability related to PCI support, in_cksum,
cpufunc.h, linkcmds etc are sufficiently challenging that just ensuring we are
able to compile for all RTEMS supported architectures will let us address many
issues.

In summary, this is just a placeholder which lets the RTEMS LibBSD code compile
on Simple Vectored architectures.