Regenerate with FreeBSD lex, yacc, and rpcgen tools.
Use the FreeBSD lex, yacc and rpcgen tools to generate the various
sources used in LibBSD. The change also removes an RTEMS specific change
added when using other hosts operating systems to generate the files.

LibBSD should use FreeBSD version of the tools to generate the source
because the FreeBSD kernel is developed using them.

The change is supported by the waf build which disables generating
the files unless enabled. I am not sure the effect with the Makefile
build support.