Add support for rc.conf(5) initialisation.
Provide user support for rc.conf(5) so a user can create a suitable
/etc/rc.conf file to initialise libbsd.

This patch by default adds basic networking support:

 cloned_interfaces
 ifconfig_'interface'
 defaultrouter
 hostname

Refer to FreeBSD documentation for examples.

Users can make a single call to have /etc/rc.conf processed, or pass a
file name to a specific configuration file or a text string with line
feeds can be passed to the scripting version of the interface.

The rc.conf support is implemented in terms of directive handlers that
are called based on a regular expression. The design allows new handlers
to be added as needed.

Line concatenation is still to be implemented.