rtemsbsd: Correctly create clocknode desc objects
When the update to FreeBSD 14/15 was performed, a clocknode description
object was manually added to clk.c. This was adequate for basic
functionality in some BSPs, but lacked the full set of definitions to
support a wider range of BSPs. This adds clknode_if.c generated from the
definitions in FreeBSD using the appropriate tool.