rtems: Update FDT support to use RTEMS_BSP_FDT
FreeBSD uses FDT to control FDT interfaces. Building with gcc15
produced errors for prototypes of functions in the code that were
not defined because we did not have `-DFDT=1` on our copmiler
command line. This change adds the global define `-DFDT=1` and
moves RTEMS specific code to be under `RTEMS_BSP_FDT` which is
detected for each BSP during configure.